Stacked bar series now uses stacked layout.
The stacked bar series uses the stack layout to compute how the
data should be stacked in the data domain. This 'layered' data
is then fed into the layout's scales & the bars are plotted.

The way the data is provided to the chart has now changed. Previously,
it was a single array of multi-value datapoint where each value
on a datapoint represented a value for a series. Now, a multi-
dimensional array is provided where the second dimension of the array
represents the data for each individual series.

The data points are assumed to have an 'x' value and a 'y' value
and the ability to set an X value accessor has been removed.

Other salient points:
- The visual test is currently setting the Y scale range max to
a fixed value.
- The crosshair is broken.