Visual test now sets domain using the stack layout.
The visual test needs to set the Y axis domain, and to do this it
needs to know about the stacked data set. The visual test uses
the stack layout to compute the layered dataset, and uses
the very top layer to calculate the maximum Y value.

We're currently using the stack layout twice in this scenario,
as we're also recalculating the stacked data when we layout.