Adds stackedData property to stackedBar.
If the user has already calculated their stacked data, say to
calculate their domains, then we should give them the opportunity
to use that data.

We still use the stacked layouts accessors to pull the values
from the data they have set, so it is their responsibility to ensure
that the data is in the format required by the stack layout.