`barSeries` uses accessors to access data values.
This allows people to pass their dataset directly into the
barSeries without manipulating it first. Note the addition of
the `getBaseline` method. This is needed because we expose
a setter for the baseline (`out`), and we need users to update
the accessor should they change the setter.

I decided to call it `getBaseline` as I felt it best described what it was
doing, but we may decide it isn't obvious enough that it is related to
`out`.