Data is now formatted into series objects.
The data is taken and transformed into an array of series objects:

[
	{
		name: ...
		data: ...
	},
	...
]

This entire series object is then passed to the stackedBar series.
Stacked bar series passes this to the stack layout, which understands
how to interpret the data as the visual test has set the 'values'
accessor.

We then use this values accessor to pull the data values from our
series objects in order to add them to our d3 selection.