StackedBar series now uses bar series to draw its bars.
The stacked series simply sets the various accessors, the baseline,
and the scales on the bar series, letting the bar series do the
work.

Note that the stacked bar series also transforms the data into an
array of data arrays, rather than an array of series objects, using
the provided `values` function. This is because the bar series
doesn't have a `values` function to transform the data itself.

The stacked bar then decorates each individual series.