Adds grouped series
The grouped series is a slight variation of the Cycle series:

The group series plots the series against the xScale (as normal)
and uses the subScale to calculate offers for these series.

This is slightly different to the Cycle series, which uses the xScale
to plot the groups, and uses the subScale to plot the position scale
within said group.