Group series now exposes offset function
The offset function returns an offset for a given group
within a set of groups. Two implementation have been
provided for this:
- One takes a interval, and will return offsets with
that value.
- One takes a fraction, which actually uses the barWidth
calculation, to work out the largest possible group size
that all groups can render, and then takes a fraction, which
it will multiple that group size by.