feat: Dynamically insert css into head or shadow-dom if missing
`d3fc-element` and `d3fc-chart` dynamically load their required css into either the document head, or else the root of the shadow dom, rather than always loading the css info the document head as a one-off.

The style element is created with an id, and we check if it exists before adding it again.

Added an example of a chart hosted in a shadow-dom.