feat: Two types of axis components, with specialisations
The default `axis` is backwards compatible (works for all scales), and has an additional `tickCenterLabel` property for positioning the label between ticks. This can be useful for continuous scales, but is less useful for ordinal scales (example included)

The alternative `axisOrdinal` will also work for all scales, and includes an optional `tickLineAlign` to control positioning of the tick line. It defaults to "center", putting the tick line in the middle of the label, but can be "left" or "right". This is more useful for ordinal scales (example included)