fix: Checks for making sure ReadMes are parsed correctly for docs
Removing the `\\n` from the regex means files that don't end with a line-feed no longer return empty content.

Adding the error check means that if a file does return empty content for any reason, it leads to a build failure rather than an undetected empty page in the docs.

re: #1247, #1248