Add files via upload
Sometimes the discontinuities follow weekly recurring patterns.
For instance, in trading, ES does not trade between 4 and 5 pm Central Time, between Monday and Thursday, and it does not trade between Fridays 5pm and Sundays 4pm Central Time.
Besides, certain symbols can have multiple non trading ranges during a day, e.g.: Corn which does not trade between 7:45 - 8:30 and 01:20 - 07:00 pm Central time.
Important to notice that these ranges should be relative to a specific timezone, which may not be the local or utc.
discontinuity.weeklypatterns.js implements the @d3fc/discontinuityProvider interface to address the above requirements.
time.js is a utility file to handle time functions. It is dependent on d3-time for floor-ing and ceil-ing dates and luxon for timezone conversion.